Funcionando

+++ entities/RetornoAPI.cs
@@ -27,15 +27,15 @@ public class HistoricalDataPrice{
     public long Data{get;set;}
 
     [JsonPropertyName("close")]
-    public float Close{get;set;}
+    public decimal Close{get;set;}
 
     [JsonPropertyName("open")]
-    public float Open{get;set;}
+    public decimal Open{get;set;}
 
     [JsonPropertyName("high")]
-    public float High{get;set;}
+    public decimal High{get;set;}
 
     [JsonPropertyName("low")]
-    public float Low{get;set;}
+    public decimal Low{get;set;}
 }
